Attach URI to Read Error
When getting a `404` or `500` error it was impossible to know which
file was causing the problem with a schema that had many external
references.  This attaches the uri that failed to the error if it's
an OpenURI error, which makes is easy to know what refrenced file
was the culprit.