Changed draft-06 url back from /draft/schema# to /draft-06/schema#
Originally, when I added draft6 support there were draft06 tests in
the common test suite, but no meta schema was available, so I assumed
the url would match the format of older
drafts (ie. http://json-schema.org/draft-06/schema#)

Then when the first version of the draft6 meta schema was released,
it's official url was http://json-schema.org/draft/schema# (although
it wasn't available on json-schema.org, it was only available on
github.com).

Now the draft6 metaschema is being hosted on json-schema.org, and the
url matches what I'd originally predicted:
http://json-schema.org/draft-06/schema#, so I've changed it back
again.

Because of that, this is largely a revert for
87e7b0aecd89ce3ff2c1dcd7dd524251cff0a464, but I've also updated the
meta schema to the latest version and updated the
`update_meta_schemas` rake task to remove the redirect to github.com
to download the meta schema.